Extract debug bundle logic
